move update results in one file
