+++ src/utils/CMakeLists.txt
@@ -1,8 +1,6 @@
 add_library(utils STATIC
             include/utils/Colors.hpp
-            include/utils/ResultFight.hpp
-            include/utils/ResultFleet.hpp
-            include/utils/ResultMerge.hpp
+            include/utils/ResultsUpdate.hpp
             include/utils/LayeredVector.hpp
             include/utils/Vec2.hpp
             include/utils/FleetInstructionType.hpp
@@ -14,11 +12,8 @@ add_library(utils STATIC
             include/utils/Probability.hpp
             include/utils/GameEventTypes.hpp
             include/utils/ResultsEvents.hpp
-            include/utils/ResultUpdate.hpp
+            include/utils/Result.hpp
             Colors.cpp
-            ResultFight.cpp
-            ResultMerge.cpp
-
 )
 target_include_directories(utils
                            PRIVATE include/utils
